chore(frontend): simplified frontend again
