--- frontend/src/app/core/mud/components/mud-input/mud-input.component.html
@@ -1,30 +1,0 @@
-<textarea
-  class="mud-input"
-  [(ngModel)]="value"
-  (keydown.enter)="submit($event)"
-  placeholder="Befehl eingeben …"
-></textarea>
-
-<!-- Nur eine der beiden Controls wird gerendert -->
-<!-- @switch (mode) {
-
-
-
-  @case ('text') {
-    <textarea
-      class="mud-input"
-      [(ngModel)]="value"
-      (keydown.enter)="submit($event)"
-      placeholder="Befehl eingeben …"
-    ></textarea>
-  }
-  @case ('password') {
-    <input
-      *ngIf="mode === 'password'"
-      class="mud-input"
-      type="password"
-      [(ngModel)]="value"
-      (keydown.enter)="submit($event)"
-    />
-  }
-} -->
