--- frontend/src/app/core/mud/components/mud-input/mud-input.component.scss
@@ -1,6 +1,0 @@
-.mud-input {
-  font-family: monospace;
-  font-size: 14px;
-  padding: 0.25rem 0.5rem;
-  resize: none; // für textarea
-}
