--- frontend/src/app/core/mud/components/mud-output/mud-output.component.html
@@ -1,1 +1,0 @@
-<div #host style="height: 100%; width: 100%;"></div>
