--- frontend/src/app/core/mud/components/mud-output/mud-output.component.scss
@@ -1,5 +1,0 @@
-:host       { display:block; height:100%; }
-.xterm      { font-size: 1rem; }               // optional
-@media (max-width:600px) {
-  .xterm { font-size: .8rem; }
-}
