+++ frontend/src/app/core/mud/index.ts
@@ -1,0 +1,2 @@
+export { MudClientComponent } from './components/mud-client/mud-client.component';
+export { MudService } from './services/mud.service';
